move cargo logic to linera_service::util
