+++ linera-service/src/cli_wrappers.rs
@@ -1,7 +1,7 @@
 // Copyright (c) Zefchain Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::config::WalletState;
+use crate::{config::WalletState, util::resolve_binary};
 use anyhow::{Context, Result};
 use async_graphql::InputType;
 use linera_base::{
@@ -11,7 +11,6 @@ use linera_base::{
     identifiers::{ChainId, MessageId, Owner},
 };
 use linera_execution::Bytecode;
-use once_cell::sync::OnceCell;
 use serde::ser::Serialize;
 use serde_json::{json, value::Value};
 use std::{
@@ -25,10 +24,7 @@ use std::{
     time::Duration,
 };
 use tempfile::{tempdir, TempDir};
-use tokio::{
-    process::{Child, Command},
-    sync::Mutex,
-};
+use tokio::process::{Child, Command};
 use tonic_health::proto::{
     health_check_response::ServingStatus, health_client::HealthClient, HealthCheckRequest,
 };
@@ -38,16 +34,12 @@ use tracing::{info, warn};
 use linera_views::common::get_table_name;
 
 /// The name of the environment variable that allows specifying additional arguments to be passed
-/// to `cargo` when starting client, server and proxy processes.
-pub const CARGO_ENV: &str = "LOCAL_NET_CARGO_PARAMS";
-
-/// The name of the environment variable that allows specifying additional arguments to be passed
 /// to the binary when starting a server.
-const SERVER_ENV: &str = "LOCAL_NET_SERVER_PARAMS";
+const SERVER_ENV: &str = "LINERA_SERVER_PARAMS";
 
 /// The name of the environment variable that allows specifying additional arguments to be passed
 /// to the node-service command of the client.
-const CLIENT_SERVICE_ENV: &str = "LOCAL_NET_CLIENT_SERVICE_PARAMS";
+const CLIENT_SERVICE_ENV: &str = "LINERA_CLIENT_SERVICE_PARAMS";
 
 #[derive(Copy, Clone)]
 pub enum Network {
@@ -1063,107 +1055,5 @@ impl NodeService {
         );
         let data = self.query_node(&query).await;
         serde_json::from_value(data["requestApplication"].clone()).unwrap()
-    }
-}
-
-#[allow(unused_mut)]
-fn detect_current_features() -> Vec<&'static str> {
-    let mut features = vec![];
-    #[cfg(feature = "benchmark")]
-    {
-        features.push("benchmark");
-    }
-    #[cfg(feature = "wasmer")]
-    {
-        features.push("wasmer");
-    }
-    #[cfg(feature = "wasmtime")]
-    {
-        features.push("wasmtime");
-    }
-    #[cfg(feature = "rocksdb")]
-    {
-        features.push("rocksdb");
-    }
-    #[cfg(feature = "scylladb")]
-    {
-        features.push("scylladb");
-    }
-    #[cfg(feature = "aws")]
-    {
-        features.push("aws");
-    }
-    features
-}
-
-async fn cargo_force_build_binary(name: &'static str, package: Option<&'static str>) -> PathBuf {
-    let package = package.unwrap_or(env!("CARGO_PKG_NAME"));
-    let mut build_command = Command::new("cargo");
-    build_command.args(["build", "-p", package]);
-    let is_release = if let Ok(var) = env::var(CARGO_ENV) {
-        let extra_args = var.split_whitespace();
-        build_command.args(extra_args.clone());
-        let extra_args: HashSet<_> = extra_args.into_iter().map(str::trim).collect();
-        extra_args.contains("-r") || extra_args.contains("--release")
-    } else {
-        false
-    };
-    // Use the same features as the current environment so that we don't rebuild as often.
-    let features = detect_current_features().join(",");
-    build_command
-        .arg("--no-default-features")
-        .arg("--features")
-        .arg(features);
-    build_command.args(["--bin", name]);
-    info!("Running compiler: {:?}", build_command);
-    assert!(build_command
-        .spawn()
-        .unwrap()
-        .wait()
-        .await
-        .unwrap()
-        .success());
-    let mut cargo_locate_command = Command::new("cargo");
-    cargo_locate_command.args(["locate-project", "--workspace", "--message-format", "plain"]);
-    let output = cargo_locate_command.output().await.unwrap().stdout;
-    let workspace_path = Path::new(std::str::from_utf8(&output).unwrap().trim())
-        .parent()
-        .unwrap();
-    if is_release {
-        workspace_path
-            .join("target/release")
-            .join(name)
-            .canonicalize()
-            .unwrap()
-    } else {
-        workspace_path
-            .join("target/debug")
-            .join(name)
-            .canonicalize()
-            .unwrap()
-    }
-}
-
-#[cfg(debug_assertions)]
-pub async fn resolve_binary(name: &'static str, package: Option<&'static str>) -> Result<PathBuf> {
-    Ok(cargo_build_binary(name, package).await)
-}
-
-#[cfg(not(debug_assertions))]
-pub async fn resolve_binary(name: &'static str, _package: Option<&'static str>) -> Result<PathBuf> {
-    crate::util::resolve_cargo_binary(name)
-}
-
-pub async fn cargo_build_binary(name: &'static str, package: Option<&'static str>) -> PathBuf {
-    type Key = (&'static str, Option<&'static str>);
-    static COMPILED_BINARIES: OnceCell<Mutex<HashMap<Key, PathBuf>>> = OnceCell::new();
-    let mut binaries = COMPILED_BINARIES.get_or_init(Default::default).lock().await;
-    match binaries.get(&(name, package)) {
-        Some(path) => path.clone(),
-        None => {
-            let path = cargo_force_build_binary(name, package).await;
-            binaries.insert((name, package), path.clone());
-            path
-        }
     }
 }
